requirements.txt
streamlit>=1.32
gspread>=6.0
google-auth>=2.28
requests>=2.31
python-dotenv>=1.0
pandas>=2.2
